Fixed words.js issue.
Resolving an issue, the last three words (the last sentence fragment) is being appended to all word counts.

Fixed it. The issue was saving to a local object stored in the Words object. Using a local object within the the Words.processWords function worked. It seems in a previous revision that's how I had initially done it and it is now back to producing the correct count.