Working on updating punctuation splitting procedures
It occurs to me that many of the errors have to do with punctuation splitting mechanics and so I've tried to account for common punctuation endings that may impact the performance. Hopefully this will lead to improved ability to handle other datasets from the two that I know work well.