That fixed it for now
The punctuation issue is currently resolved. However, the punctuation that is left in the text is still not included in the span and could likely be fixed now that the original sentence object contains the original punctuation that belongs to the sentence. As a backup, in case that original punctuation does not make it to the final span-decoupler interaction; we can always just try to use a random assortment of punctuation in or replace or search algorithm to ensure we get it right.